Remove commented-out DEBUG_STORE_DATA calls.
Also remove the SILK_DEBUG_STORE_CLOSE_FILES flush call from
opus_demo.

This is debugging code which is no longer used, but defining
the symbols for SILK_DEBUG_STORE_CLOSE_FILES and calling it
from opus_demo causes linking problems on Microsoft Visual Studio
where we have strict controls on public symbols and want to
test the compiled DLL.

Since the code isn't in active use, it's better to remove it
to avoid clutter and address the linking issue.